feat: add input validation util class
- Added input validation util class

Changed location of the source code that user's input validation

Before:
 1. isCarName5Characters method in PlayRacingCar class
 2. isValidNumber method in PlayRacingCar class

After
 1. isCarName5Characters method in InputValidationUtil class
 2. isValidNumber method in InputValidationUtil class
